requirements.txt
django
google-generativeai
pillow
sentence-transformers
psycopg2-binary
gunicorn
